fixed the submit functionality, added put and delete functionality but not finish yet
