+++ friends/src/index.css
@@ -3,3 +3,8 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
+
+.card-container {
+  justify-content: space-around;
+}
+
